requirements.txt
pytesseract
Pillow
PyYAML
PyMuPDF
opencv-python-headless
numpy
transformers
timm
pandas
openpyxl
openai
streamlit
scikit-learn

